requirements.txt
discord.py
python-dotenv
psycopg2-binary
openai
flask
werkzeug